% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic_Estimation.R
\name{Logistic_Estimation}
\alias{Logistic_Estimation}
\title{Logistic Regression Estimation}
\usage{
Logistic_Estimation(X, y)
}
\arguments{
\item{X}{A numeric matrix where each row is an observation and each column is a predictor variable.}

\item{y}{A numeric vector of binary response variables (0 or 1) corresponding to each row of X.}
}
\value{
A list containing various components:
  - Coefficients: A data frame of logistic regression coefficients, standard errors, Z-scores, and p-values.
  - beta: The final estimated coefficients.
  - y_pred: Predicted values based on the logistic model.
  - times: The number of iterations taken in the Newton-Raphson process.
  - J: The Fisher Information Matrix.
  - J_inv: The inverse of the Fisher Information Matrix.
  - delta: The final delta value in the iteration process.
  - confusion_matrix: A confusion matrix of the model.
}
\description{
Performs logistic regression estimation using the Newton-Raphson method. This function estimates coefficients for logistic regression and computes various statistics, including a confusion matrix, standard errors, Z-scores, and p-values.
}
\examples{
X <- matrix(rnorm(100), ncol=5)
y <- rbinom(20, 1, 0.5)
result <- Logistic_Estimation(X, y)
print(result$Coefficients)
print(result$confusion_matrix)

X2 <- matrix(rnorm(20), ncol=2, dimnames = list(NULL, c("Variable1", "Variable2")))
y2 <- rbinom(10, 1, 0.5)
result2 <- Logistic_Estimation(X2, y2)
print(result2$Coefficients)
print(result2$confusion_matrix)
}
