% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic_Estimation.R
\name{Logistic_Estimation}
\alias{Logistic_Estimation}
\title{Perform Logistic Regression Estimation}
\usage{
Logistic_Estimation(X, y, print_result = FALSE)
}
\arguments{
\item{X}{A numeric matrix where each row represents an observation and each column
represents a predictor variable.}

\item{y}{A numeric vector of binary response variables (0 or 1) corresponding
to each row of X.}

\item{print_result}{Logical, indicating whether to print the results. Defaults to FALSE.}
}
\value{
A list containing various components:
  - Coefficients: A data frame of logistic regression coefficients, standard errors,
    Z-scores, and p-values.
  - beta: The final estimated coefficients.
  - y_pred: Predicted values based on the logistic model.
  - times: The number of iterations taken in the Newton-Raphson process.
  - J: The Fisher Information Matrix.
  - J_inv: The inverse of the Fisher Information Matrix.
  - delta: The final delta value in the iteration process.
  - confusion_matrix: A confusion matrix of the model.
}
\description{
This function performs logistic regression estimation using the
Newton-Raphson method. It estimates coefficients, calculates standard errors,
Z-scores, p-values, and generates a confusion matrix. The function also allows
printing the results.
}
\examples{
set.seed(1)
X = matrix(rnorm(100), ncol = 5)
y =rbinom(20, 1, 0.5)
result =Logistic_Estimation(X, y)
print(result$Coefficients)
print(result$confusion_matrix)


X =matrix(rnorm(50), ncol = 5)
y =rbinom(10, 1, 0.5)
result =Logistic_Estimation(X, y, print_result = TRUE)
print(result$Coefficients)
print(result$confusion_matrix)
}
