% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logit.R
\name{Logit}
\alias{Logit}
\title{Calculate Logit Values}
\usage{
Logit(X, beta)
}
\arguments{
\item{X}{A numeric matrix of predictors where each row represents an observation and each column represents a predictor variable.}

\item{beta}{A numeric vector of coefficients corresponding to the predictor variables.}
}
\value{
A numeric vector of logit values. When handling extreme values, the function adjusts the computation to maintain numerical stability.
}
\description{
This function calculates logit values for a given set of predictors and coefficients. It is designed to handle extreme values in the exponentiation to avoid computational errors.
}
\examples{
X <- matrix(rnorm(20), ncol=2)
beta <- runif(2)
logit_values <- Logit(X, beta)
}
